added wallet to GetUsersList with admin and supervisor users filter

+++ Accounting/Models/UserRole.cs
@@ -12,5 +12,5 @@ public partial class UserRole
 
     public long RegUserId { get; set; }
 
-    public short RoleId { get; set; }
+    public int RoleId { get; set; }
 }
